Cambios hechos para arreglar codigo, falta revisar por bugs y ya esta funcional

+++ src/main/java/org/JavaCar/Roda.java
@@ -2,9 +2,9 @@ package org.JavaCar;
 
 public class Roda {
     private String marca;
-    private double diametre;
+    private int diametre;
 
-    public Roda(String marca, double diametre) {
+    public Roda(String marca, int diametre) {
         this.marca = marca;
         this.diametre = diametre;
     }
